src/fraud_doc_detection/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pymupdf==1.24.5
pdfplumber==0.11.1
Pillow==10.3.0
numpy==1.26.4
scikit-learn==1.5.0
python-magic==0.4.27
cryptography==42.0.8
python-dateutil==2.9.0.post0
aiofiles==23.2.1
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.3.1
google-generativeai==0.7.2
python-dotenv==1.0.1
openai>=1.30.0
